pyproject.toml
[tool.poetry]
name = "fastapi-shop-api"
version = "0.1.0"
description = ""
authors = ["AlekseyCherniaev <cherniaevaleksey2304@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["uvicorn"], version = "^0.111.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.31"}
pytest = "^8.2.2"
pydantic-settings = "^2.3.4"
asyncpg = "^0.29.0"
alembic = {extras = ["async"], version = "^1.13.2"}
bcrypt = "^4.2.0"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
itsdangerous = "^2.2.0"
authlib = "^1.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
